Changed the html for album deletion to fix an error, Exceptions are handled for different kinds of malfunctions, for example if your try to get and non-existant artist

+++ src/main/java/com/myaudiolibrary/web/controller/AlbumController.java
@@ -12,6 +12,8 @@ import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.view.RedirectView;
 
+import javax.persistence.EntityNotFoundException;
+
 //@CrossOrigin
 //@RestController
 @Controller
@@ -24,6 +26,14 @@ public class AlbumController {
     @RequestMapping(method = RequestMethod.POST, consumes = MediaType.APPLICATION_FORM_URLENCODED_VALUE)
     public RedirectView registerAlbum(final ModelMap albumMap, Album album, Artist artist)
     {
+        if(album.getTitle().isEmpty()){
+            throw new IllegalArgumentException("Vous ne pouvez pas créer d'albums sans nom !");
+        }
+
+        if(album.getArtist() == null){
+            throw new IllegalArgumentException("Vous ne pouvez pas ajouter un album sans artiste !");
+        }
+
         albumRepository.save(album);
         return new RedirectView("/artists/"+artist.getId());
 
@@ -31,6 +41,9 @@ public class AlbumController {
 
     @RequestMapping(method = RequestMethod.GET, value = "/{albumId}")
     public RedirectView deleteAlbum(@PathVariable("albumId") Integer albumId, final ModelMap albumMap, Artist artist){
+        if(!(albumRepository.existsById(albumId))){
+            throw new EntityNotFoundException("L'album d'id "+albumId+" N'existe pas !");
+        }
         albumRepository.deleteById(albumId);
         return new RedirectView("/artists/"+artist.getId());
     }
