+++ src/main/java/com/myaudiolibrary/web/controller/ArtistController.java
@@ -40,7 +40,14 @@ public class ArtistController {
             final ModelMap modelArtist,
             @PathVariable("idArtist") Integer idArtist
     ){
+        if(idArtist <= 0){
+            throw new IllegalArgumentException("L'id de l'artiste doit être un chiffre non nul");
+        }
+
         Optional<Artist> artistOptional = artistRepository.findById(idArtist);
+        if(artistOptional.isEmpty()){
+            throw new EntityNotFoundException("L'artiste d'id "+idArtist+" n'existe pas !");
+        }
         modelArtist.put("Artist", artistOptional.get());
         modelArtist.put("albumToCreate", new Album());
         return "detailArtist";
@@ -56,7 +63,7 @@ public class ArtistController {
             @RequestParam(value = "sortProperty", defaultValue = "name") String sortProperty,
             @RequestParam(value = "sortDirection", defaultValue = "ASC") String sortDirection
     ){
-        /*if(page < 0){
+        if(page < 0){
             throw new IllegalArgumentException("La valeur Page ne peut pas être négative !");
         }
         if(size <= 0 || size >= 50){
@@ -64,7 +71,7 @@ public class ArtistController {
         }
         if(!("ASC".equalsIgnoreCase(sortDirection)) && !("DESC".equalsIgnoreCase(sortDirection))){
             throw new IllegalArgumentException("Le paramètre sortDirection doit valoir soit ASC soit DESC");
-        }*/
+        }
 
         PageRequest pageRequest =  PageRequest.of(page, size, Sort.Direction.fromString(sortDirection), sortProperty);
         Page<Artist> artistList= artistRepository.findAllByNameContaining(name, pageRequest);
@@ -92,7 +99,7 @@ public class ArtistController {
             @RequestParam(value = "sortProperty", defaultValue = "name") String sortProperty,
             @RequestParam(value = "sortDirection", defaultValue = "ASC") String sortDirection
     ){
-        /*if(page < 0){
+        if(page < 0){
             throw new IllegalArgumentException("La valeur Page ne peut pas être négative !");
         }
         if(size <= 0 || size >= 50){
@@ -100,7 +107,7 @@ public class ArtistController {
         }
         if(!("ASC".equalsIgnoreCase(sortDirection)) && !("DESC".equalsIgnoreCase(sortDirection))){
             throw new IllegalArgumentException("Le paramètre sortDirection doit valoir soit ASC soit DESC");
-        }*/
+        }
 
         PageRequest pageRequest =  PageRequest.of(page, size, Sort.Direction.fromString(sortDirection), sortProperty);
         Page<Artist> artistList= artistRepository.findAll( pageRequest);
@@ -128,10 +135,16 @@ public class ArtistController {
 
     @RequestMapping(method = RequestMethod.POST, value = "/registerArtist", consumes = MediaType.APPLICATION_FORM_URLENCODED_VALUE)
     public RedirectView registerArtist(Artist artist, final ModelMap artistMap){
+        if(artistRepository.existsByName(artist.getName())){
+            throw new EntityExistsException("L'artiste que vous essayer d'ajouter existe déjà !");
+        }
         return saveArtist(artist, artistMap);
     }
 
     private RedirectView saveArtist(Artist artist, ModelMap artistMap){
+        if(artistRepository.existsByName(artist.getName())){
+            throw new EntityExistsException("L'artiste que vous essayer d'ajouter existe déjà !");
+        }
         artist = artistRepository.save(artist);
         artistMap.put("Artist", artist);
         return new RedirectView("/artists/"+ artist.getId());
@@ -139,6 +152,16 @@ public class ArtistController {
 
     @RequestMapping(method = RequestMethod.GET, value = "/{idArtist}/delete")
     public RedirectView deleteArtist(@PathVariable("idArtist") Integer idArtist, final ModelMap artistMap){
+
+        Optional<Artist> artistCheck = artistRepository.findById(idArtist);
+        if (artistCheck.isEmpty()){
+            throw new EntityNotFoundException("L'artiste que vous essayer de supprimer n'existe pas !");
+        }
+
+        if(artistCheck.get().getAlbums() == null || !(artistCheck.get().getAlbums().isEmpty())){
+            throw new IllegalArgumentException("L'artiste possède encore des albums, vous ne pouvez pas le supprimer !");
+        }
+
         artistRepository.deleteById(idArtist);
         return new RedirectView("/artists");
     }
